# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Globapay Platform API
  description: Multi-tenant payments orchestration API with contract-first design
  version: 0.1.0
  contact:
    name: Globapay Platform Team
    url: https://globapay.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.globapay.com
    description: Production server
security:
  - bearerAuth: []
  - apiKey: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: Unique key for request idempotency
      required: false
      schema:
        type: string
        maxLength: 255
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: Problem type identifier
        title:
          type: string
          description: Human-readable problem summary
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable problem details
        instance:
          type: string
          format: uri
          description: Problem instance identifier
      required: [type, title, status]
...
    WebhookIncomingRequest:
      type: object
      properties:
        type:
          type: string
        data:
          type: object
        signature:
          type: string
        timestamp:
          type: string
          format: date-time
      required: [type, data, signature, timestamp]
    WebhookReplayRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
      required: [reason]
...
paths:
...
  /webhooks/fraud:
    post:
      summary: Receive webhook from fraud detection service
      operationId: receiveFraudWebhook
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookIncomingRequest'
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [processed]
                required: [status]
        '400':
          $ref: '#/components/responses/BadRequest'
  /webhooks/kyb:
    post:
      summary: Receive webhook from KYB verification service
      operationId: receiveKYBWebhook
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookIncomingRequest'
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [processed]
                required: [status]
        '400':
          $ref: '#/components/responses/BadRequest'

  # Admin - Webhook Management
  /webhooks/{id}/replay:
    post:
      summary: Replay a webhook event
      operationId: replayWebhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookReplayRequest'
      responses:
        '200':
          description: Webhook replay initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [queued]
                required: [status]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Reports
...